refactor(app): migrate useHomePipettes to local-resources
