+++ app/src/local-resources/instruments/hooks.ts
@@ -6,6 +6,7 @@ import {
   GRIPPER_MODELS,
 } from '@opentrons/shared-data'
 import { useIsOEMMode } from '/app/resources/robot-settings/hooks'
+import { useRobotControlCommands } from '/app/resources/maintenance_runs'
 
 import type {
   GripperModel,
@@ -14,7 +15,12 @@ import type {
   PipetteName,
   PipetteNameSpecs,
   PipetteV2Specs,
+  CreateCommand,
 } from '@opentrons/shared-data'
+import type {
+  UseRobotControlCommandsProps,
+  UseRobotControlCommandsResult,
+} from '/app/resources/maintenance_runs'
 
 export function usePipetteNameSpecs(
   name: PipetteName
@@ -22,7 +28,9 @@ export function usePipetteNameSpecs(
   const isOEMMode = useIsOEMMode()
   const pipetteNameSpecs = getPipetteNameSpecs(name)
 
-  if (pipetteNameSpecs == null) return null
+  if (pipetteNameSpecs == null) {
+    return null
+  }
 
   const brandedDisplayName = pipetteNameSpecs.displayName
   const anonymizedDisplayName = pipetteNameSpecs.displayName.replace(
@@ -43,7 +51,9 @@ export function usePipetteModelSpecs(
     modelSpecificFields?.name as PipetteName
   )
 
-  if (modelSpecificFields == null || pipetteNameSpecs == null) return null
+  if (modelSpecificFields == null || pipetteNameSpecs == null) {
+    return null
+  }
 
   return { ...modelSpecificFields, displayName: pipetteNameSpecs.displayName }
 }
@@ -54,7 +64,9 @@ export function usePipetteSpecsV2(
   const isOEMMode = useIsOEMMode()
   const pipetteSpecs = getPipetteSpecsV2(name)
 
-  if (pipetteSpecs == null) return null
+  if (pipetteSpecs == null) {
+    return null
+  }
 
   const brandedDisplayName = pipetteSpecs.displayName
   const anonymizedDisplayName = pipetteSpecs.displayName.replace('Flex ', '')
@@ -77,4 +89,35 @@ export function useGripperDisplayName(gripperModel: GripperModel): string {
   const anonymizedDisplayName = brandedDisplayName.replace('Flex ', '')
 
   return isOEMMode ? anonymizedDisplayName : brandedDisplayName
+}
+
+export interface UseHomePipettesResult {
+  isHoming: UseRobotControlCommandsResult['isExecuting']
+  homePipettes: UseRobotControlCommandsResult['executeCommands']
+}
+
+export type UseHomePipettesProps = Pick<
+  UseRobotControlCommandsProps,
+  'pipetteInfo' | 'onSettled'
+>
+
+// Home pipettes except for plungers.
+export function useHomePipettes(
+  props: UseHomePipettesProps
+): UseHomePipettesResult {
+  const { executeCommands, isExecuting } = useRobotControlCommands({
+    ...props,
+    commands: [HOME_EXCEPT_PLUNGERS],
+    continuePastCommandFailure: true,
+  })
+
+  return {
+    isHoming: isExecuting,
+    homePipettes: executeCommands,
+  }
+}
+
+const HOME_EXCEPT_PLUNGERS: CreateCommand = {
+  commandType: 'home' as const,
+  params: { axes: ['leftZ', 'rightZ', 'x', 'y'] },
 }
