+++ app/src/organisms/DropTipWizardFlows/TipsAttachedModal.tsx
@@ -14,11 +14,12 @@ import { FLEX_ROBOT_TYPE } from '@opentrons/shared-data'
 import { SmallButton } from '/app/atoms/buttons'
 import { OddModal } from '/app/molecules/OddModal'
 import { DropTipWizardFlows, useDropTipWizardFlows } from '.'
-import { useHomePipettes } from './hooks'
+import { useHomePipettes } from '/app/local-resources/instruments'
 
 import type { HostConfig } from '@opentrons/api-client'
 import type { OddModalHeaderBaseProps } from '/app/molecules/OddModal/types'
-import type { UseHomePipettesProps, PipetteWithTip } from './hooks'
+import type { UseHomePipettesProps } from '/app/local-resources/instruments'
+import type { PipetteWithTip } from './hooks'
 import type { PipetteDetails } from '/app/resources/maintenance_runs'
 
 type TipsAttachedModalProps = Pick<UseHomePipettesProps, 'onSettled'> & {
