+++ app/src/organisms/DropTipWizardFlows/index.ts
@@ -1,10 +1,6 @@
 export * from './DropTipWizardFlows'
-export { useTipAttachmentStatus, useHomePipettes } from './hooks'
+export { useTipAttachmentStatus } from './hooks'
 export * from './TipsAttachedModal'
 
-export type {
-  UseHomePipettesProps,
-  TipAttachmentStatusResult,
-  PipetteWithTip,
-} from './hooks'
+export type { TipAttachmentStatusResult, PipetteWithTip } from './hooks'
 export type { FixitCommandTypeUtils } from './types'
